Cleanup + Plotting + Comments
