+++ Demo_SupplyChain.py
@@ -4,11 +4,6 @@ import numpy as np
 from VISolver.Domains.SupplyChain import (
     SupplyChain, CreateRandomNetwork)
 
-# from VISolver.Solvers.Euler import Euler
-# from VISolver.Solvers.Extragradient import EG
-# from VISolver.Solvers.AcceleratedGradient import AG
-# from VISolver.Solvers.HeunEuler import HeunEuler
-# from VISolver.Solvers.AdamsBashforthEuler import ABEuler
 from VISolver.Solvers.CashKarp import CashKarp
 
 from VISolver.Projection import BoxProjection
@@ -27,11 +22,6 @@ def Demo():
     Domain = SupplyChain(Network=Network,alpha=2)
 
     # Set Method
-    # Method = Euler(Domain=Domain,P=BoxProjection(lo=0))
-    # Method = EG(Domain=Domain,P=BoxProjection(lo=0))
-    # Method = AG(Domain=Domain,P=BoxProjection(lo=0))
-    # Method = HeunEuler(Domain=Domain,P=BoxProjection(lo=0),Delta0=1e-2)
-    # Method = ABEuler(Domain=Domain,P=BoxProjection(lo=0),Delta0=1e-2)
     Method = CashKarp(Domain=Domain,P=BoxProjection(lo=0),Delta0=1e-5)
 
     # Initialize Starting Point
@@ -45,7 +35,6 @@ def Demo():
 
     # Set Options
     Init = Initialization(Step=-1e-10)
-    # Init = Initialization(Step=-0.1)
     Term = Termination(MaxIter=25000,Tols=[(Domain.gap_rplus,1e-3*gap_0)])
     Repo = Reporting(Requests=[Domain.gap_rplus, 'Step', 'F Evaluations',
                                'Projections','Data'])
